backend/requirements.txt
fastapi
uvicorn
python-multipart
twelvelabs
yt-dlp
deep-translator
python-dotenv
